replace nextcloud and item classes
