+++ src/handler/eventHandler.h
@@ -11,7 +11,6 @@
 
 #include "contextMenu.h"
 #include "mainMenu.h"
-#include "nextcloud.h"
 #include "webDAVView.h"
 #include "loginView.h"
 #include "sqliteConnector.h"
@@ -19,6 +18,7 @@
 #include <memory>
 
 const std::string CONFIG_FOLDER = "/mnt/ext1/system/config/nextcloud";
+const std::string CONFIG_PATH = CONFIG_FOLDER + "/nextcloud.cfg";
 //TODO use folder of nextcloud conifg temp
 const std::string DB_PATH = CONFIG_FOLDER + "/data.db";
